requirements.txt
google-generativeai
transformers
numpy
tf-keras
sentence-transformers
faiss-cpu
fastai